+++ vfrecovery/command_line_interface/virtualfleet_recovery.py
@@ -2,14 +2,12 @@ import click
 
 from vfrecovery.command_line_interface.group_describe import cli_group_describe
 from vfrecovery.command_line_interface.group_predict import cli_group_predict
-from vfrecovery.command_line_interface.group_plot import cli_group_plot
 
 @click.command(
     cls=click.CommandCollection,
     sources=[
         cli_group_describe,
         cli_group_predict,
-        cli_group_plot,
     ],
     context_settings=dict(help_option_names=["-h", "--help"]),
 )
