+++ vfrecovery/core/__init__.py
@@ -1,4 +1,4 @@
 # from deployment_plan import setup_deployment_plan
 from .trajfile_handler import Trajectories
-from .run_handler import RunAnalyser
-# from predict import predict_function
+from .analysis_handler import RunAnalyser
+from .simulation_handler import Simulation
